+++ frontend/src/domains/admin/ManageRoomModal/hooks/useDeleteOrganization.ts
